mouse_pos should be ints

+++ mcio_remote/network.py
@@ -124,7 +124,7 @@ class ActionPacket:
     )  # List of (button, action) pairs
 
     # List of (x, y) pairs. Using a list for consistency
-    mouse_pos: List[Tuple[float, float]] = field(default_factory=list)
+    mouse_pos: List[Tuple[int, int]] = field(default_factory=list)
 
     def pack(self) -> bytes:
         pkt_dict = asdict(self)
